Refactored code to AsBaggageHeader extension for consistency
