+++ src/Sentry.OpenTelemetry/OpenTelemetryExtensions.cs
@@ -9,4 +9,11 @@ internal static class OpenTelemetryExtensions
     public static SentryId AsSentryId(this ActivityTraceId id) => SentryId.Parse(id.ToHexString());
 
     public static ActivityTraceId AsActivityTraceId(this SentryId id) => ActivityTraceId.CreateFromString(id.ToString().AsSpan());
+
+    public static BaggageHeader AsBaggageHeader(this IEnumerable<KeyValuePair<string, string?>> baggage, bool useSentryPrefix = false) =>
+        BaggageHeader.Create(
+            baggage.Where(member => member.Value != null)
+                        .Select(kvp => (KeyValuePair<string, string>)kvp!),
+            useSentryPrefix
+            );
 }
